dashboard/requirements.txt
astroid==1.6.5
matplotlib==3.1.2
autopep8==1.3.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
seaborn==0.9.0
click==6.7
cycler==0.10.0
dash==0.21.1
dash-core-components==0.23.0
dash-html-components==0.11.0
dash-renderer==0.13.0
dash-table-experiments==0.6.0
decorator==4.3.0
dill==0.2.7.1
Flask==1.0.2
Flask-Compress==1.4.0
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.5.0
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
httplib2==0.11.3
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nbformat==4.4.0
numpy==1.14.5
oauth2client==3.0.0
pandas==0.25.3
pkg-resources==0.0.0
plotly==2.7.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.0
psutil==5.4.6
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pylint==1.9.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
rsa==3.4.2
scipy==1.1.0
six==1.11.0
traitlets==4.3.2
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
